\name{browsegWidgetsFiles}
\alias{browsegWidgetsFiles}
\title{Simple GUI to browse gWidgets examples}
\usage{
  browsegWidgetsFiles(toolkit = "RGtk2")
}
\arguments{
  \item{toolkit}{Which toolkit to use (RGtk2, tcltk or Qt)}
}
\value{
  NULL
}
\description{
  Simple GUI to browse gWidgets examples
}

