\name{day.of.week}
\alias{day.of.week}
\title{Day of week for a given year, month day}
\usage{
  day.of.week(year, month, day)
}
\arguments{
  \item{year}{year}

  \item{month}{month as integer 1 to 12}

  \item{day}{day of month}
}
\value{
  day of week in 0:6
}
\description{
  from chron with slight change to arguments
}

