\name{ProgGUIinR-package}
\alias{ProgGUIinR-package}
\alias{ProgGUIinR}
\docType{package}
\title{
Support package for Programming GUIs in R
}
\description{
Examples, Sample code, appendices and functions for the text Programming GUIs inR
}
\details{
This package contains the code from the book Programming GUIs in R. 	
}
\author{
John Verzani and Michael Lawrence

Maintainer: John Verzani <verzani@math.csi.cuny.edu>

}
\references{

}

\keyword{ package }
\seealso{
}
\examples{

}
